+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changes
+#### Improvements
+
+- Updates `change` field on `LogEntry` model to be a json rather than a text field ([#407](https://github.com/jazzband/django-auditlog/pull/407))
 
 ## Next Release